SystemGesture - Rename files to match naming convention
